Move toolbar back out of secretlistheader
