+++ ui/app/controllers/vault/cluster/secrets/backend/configuration.js
@@ -1,0 +1,9 @@
+import { computed } from '@ember/object';
+import Controller from '@ember/controller';
+
+export default Controller.extend({
+  isConfigurable: computed('model.type', function() {
+    const configurableEngines = ['aws', 'ssh', 'pki'];
+    return configurableEngines.includes(this.get('model.type'));
+  }),
+});
