+++ ui/app/templates/components/secret-list-header.hbs
@@ -75,60 +75,3 @@
     </div>
   {{/if}}
 {{/with}}
-
-{{#if (or hasItems (or isConfigurable (not isConfigurableTab)))}}
-  <Toolbar>
-    {{#if hasItems}}
-      <ToolbarFilters>
-        {{navigate-input
-          enterpriseProduct="vault"
-          filterFocusDidChange=(action "setFilterFocus")
-          filterDidChange=(action "setFilter")
-          filter=filter
-          filterMatchesKey=filterMatchesKey
-          firstPartialMatch=firstPartialMatch
-          baseKey=(get baseKey "id")
-          shouldNavigateTree=options.navigateTree
-          placeholder=options.searchPlaceholder
-          mode=(if (eq tab 'certs') 'secrets-cert' 'secrets')
-          data-test-nav-input=true
-        }}
-        {{#if filterFocused}}
-          {{#if filterMatchesKey}}
-            {{#unless filterIsFolder}}
-              <p class="input-hint">
-                <kbd>Enter</kbd> to view {{filter}}
-              </p>
-            {{/unless}}
-          {{/if}}
-          {{#if firstPartialMatch}}
-            <p class="input-hint">
-              <kbd>Tab</kbd> to autocomplete
-            </p>
-          {{/if}}
-        {{/if}}
-      </ToolbarFilters>
-    {{/if}}
-
-    <ToolbarActions>
-      {{#if (and isConfigurable isConfigurableTab)}}
-        <ToolbarLink
-          @params={{array 'vault.cluster.settings.configure-secret-backend' model.id}}
-          @data-test-secret-backend-configure=true
-        >
-          Configure
-        </ToolbarLink>
-      {{else}}
-        <ToolbarSecretLink
-          @secret=''
-          @mode="create"
-          @type="add"
-          @queryParams={{query-params initialKey=(or filter baseKey.id)}}
-          @data-test-secret-create=true
-        >
-          {{options.create}}
-        </ToolbarSecretLink>
-      {{/if}}
-    </ToolbarActions>
-  </Toolbar>
-{{/if}}
