+++ ui/app/templates/vault/cluster/secrets/backend/configuration.hbs
@@ -4,6 +4,19 @@
   @isConfigure=true
 />
 
+{{#if isConfigurable}}
+  <Toolbar>
+    <ToolbarActions>
+      <ToolbarLink
+        @params={{array 'vault.cluster.settings.configure-secret-backend' model.id}}
+        @data-test-secret-backend-configure=true
+      >
+        Configure
+      </ToolbarLink>
+    </ToolbarActions>
+  </Toolbar>
+{{/if}}
+
 <div class="box is-fullwidth is-sideless is-paddingless is-marginless">
   {{#each model.attrs as |attr|}}
     {{#if (eq attr.type "object")}}
