+++ ui/app/templates/vault/cluster/secrets/backend/list.hbs
@@ -1,12 +1,59 @@
+<SecretListHeader
+  @isCertTab={{eq tab "certs"}}
+  @model={{backendModel}}
+  @baseKey={{baseKey}}
+  @backendCrumb={{backendCrumb}}
+  @filter={{filter}}
+/>
+
 {{#with (options-for-backend backendType tab) as |options|}}
-  <SecretListHeader
-    @isCertTab={{eq tab "certs"}}
-    @model={{backendModel}}
-    @baseKey={{baseKey}}
-    @backendCrumb={{backendCrumb}}
-    @filter={{filter}}
-    @options={{options}}
-  />
+  {{#if (or model.meta.total (not isConfigurableTab))}}
+    <Toolbar>
+      {{#if model.meta.total}}
+        <ToolbarFilters>
+          {{navigate-input
+            enterpriseProduct="vault"
+            filterFocusDidChange=(action "setFilterFocus")
+            filterDidChange=(action "setFilter")
+            filter=filter
+            filterMatchesKey=filterMatchesKey
+            firstPartialMatch=firstPartialMatch
+            baseKey=(get baseKey "id")
+            shouldNavigateTree=options.navigateTree
+            placeholder=options.searchPlaceholder
+            mode=(if (eq tab 'certs') 'secrets-cert' 'secrets')
+            data-test-nav-input=true
+          }}
+          {{#if filterFocused}}
+            {{#if filterMatchesKey}}
+              {{#unless filterIsFolder}}
+                <p class="input-hint">
+                  <kbd>Enter</kbd> to view {{filter}}
+                </p>
+              {{/unless}}
+            {{/if}}
+            {{#if firstPartialMatch}}
+              <p class="input-hint">
+                <kbd>Tab</kbd> to autocomplete
+              </p>
+            {{/if}}
+          {{/if}}
+        </ToolbarFilters>
+      {{/if}}
+
+      <ToolbarActions>
+        <ToolbarSecretLink
+          @secret=''
+          @mode="create"
+          @type="add"
+          @queryParams={{query-params initialKey=(or filter baseKey.id)}}
+          @data-test-secret-create=true
+        >
+          {{options.create}}
+        </ToolbarSecretLink>
+      </ToolbarActions>
+    </Toolbar>
+  {{/if}}
 
   {{#if model.meta.total}}
     {{#each model as |item|}}
